make news file
