+++ iCol/ViewControllers/HabitDetailViewController.swift
@@ -155,56 +155,6 @@ extension ChallengeProgressView: UICollectionViewDelegateFlowLayout, UICollectio
     
 }
 
-class DayCell: UICollectionViewCell {
-    
-    static let reuseIdentifier = "DayCell"
-    
-    let dateLabel = UILabel()
-    let monthLabel = UILabel()
-    let tickButton = UIButton(type: .system)
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        
-        let containerView = UIView()
-        containerView.backgroundColor = #colorLiteral(red: 0.8850145936, green: 0.9960160851, blue: 0.9677742124, alpha: 1)
-        containerView.layer.cornerRadius = 8
-
-        addSubview(containerView)
-        
-        dateLabel.font = .preferredFont(forTextStyle: .body)
-        dateLabel.font = .boldSystemFont(ofSize: 17)
-        dateLabel.text = "23"
-        monthLabel.font = .preferredFont(forTextStyle: .footnote)
-        monthLabel.text = "Oct"
-        
-        tickButton.setImage(UIImage(systemName: "person.circle"), for: .normal)
-        containerView.addSubview(tickButton)
-        
-        let stackView = UIStackView(arrangedSubviews: [dateLabel, monthLabel])
-        stackView.axis = .vertical
-        stackView.alignment = .center
-        containerView.addSubview(stackView)
-        
-        containerView.setConstraint(topAnchor: topAnchor, topAnchorConstant: 0,
-                                    leadingAnchor: leadingAnchor, leadingAnchorConstant: 0,
-                                    trailingAnchor: trailingAnchor, trailingAnchorConstant: 0)
-        containerView.heightAnchor.constraint(equalToConstant: 64).isActive = true
-        
-        stackView.setConstraint(centerXAnchor: containerView.centerXAnchor, centerXAnchorConstant: 0,
-                                centerYAnchor: containerView.centerYAnchor, centerYAnchorConstant: 0)
-        
-        tickButton.setConstraint(topAnchor: stackView.bottomAnchor, topAnchorConstant: 32,
-                                 bottomAnchor: bottomAnchor, bottomAnchorConstant: 0,
-                                 centerXAnchor: containerView.centerXAnchor, centerXAnchorConstant: 0)
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-}
-
 class ChallengeOverview: UIView {
     
     let percentageLabel = UILabel()
@@ -311,43 +261,3 @@ class ChallengeOverview: UIView {
 
 }
 
-class ProgressBarView: UIView {
-
-    private let progressView = UIView()
-    private let progressBackground = UIView()
-
-    var percentage = CGFloat() {
-        didSet {
-            progressView.widthAnchor.constraint(equalTo: progressBackground.widthAnchor, multiplier: percentage).isActive = true
-        }
-    }
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-
-        addSubview(progressBackground)
-        addSubview(progressView)
-
-        progressView.backgroundColor = #colorLiteral(red: 0.176264137, green: 0.3899727762, blue: 0.334807694, alpha: 1)
-        progressView.layer.cornerRadius = 8
-        progressBackground.backgroundColor = #colorLiteral(red: 0.7685508132, green: 0.768681109, blue: 0.7685337067, alpha: 1)
-        progressBackground.layer.cornerRadius = 8
-
-        progressBackground.setConstraint(topAnchor: topAnchor, topAnchorConstant: 0,
-                                         bottomAnchor: bottomAnchor,
-                                         leadingAnchor: leadingAnchor, leadingAnchorConstant: 0,
-                                         trailingAnchor: trailingAnchor, trailingAnchorConstant: 0)
-        progressBackground.heightAnchor.constraint(equalToConstant: 30).isActive = true
-        progressBackground.widthAnchor.constraint(equalTo: widthAnchor).isActive = true
-
-        progressView.setConstraint(topAnchor: topAnchor, topAnchorConstant: 0,
-                                   bottomAnchor: bottomAnchor,
-                                   leadingAnchor: leadingAnchor, leadingAnchorConstant: 0)
-        progressView.heightAnchor.constraint(equalToConstant: 30).isActive = true
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-}
